re-factored the snackbar into own jsx file, removed bug from useEffect in Settings, and condensed all of the change handlers into one change handler
