+++ client/src/components/Snackbar/Snackbar.jsx
@@ -1,0 +1,48 @@
+import React from 'react';
+import IconButton from "@material-ui/core/IconButton";
+import CloseIcon from "@material-ui/icons/Close";
+import Snackbar from "@material-ui/core/Snackbar";
+import {SnackbarContent} from "@material-ui/core";
+
+function Snack(props) {
+  const handleClose = (event, reason) => {
+    if (reason === "clickaway") {
+      return;
+    }
+
+    props.setOpen(false);
+  };
+
+  return (
+    <Snackbar
+      anchorOrigin={{
+        vertical: "bottom",
+        horizontal: "center",
+      }}
+      open={props.open}
+      autoHideDuration={3000}
+      onClose={handleClose}
+      action={
+        <React.Fragment>
+          <IconButton
+            size="small"
+            aria-label="close"
+            color="secondary"
+            onClick={handleClose}
+          >
+            <CloseIcon fontSize="small" />
+          </IconButton>
+        </React.Fragment>
+      }
+    >
+      <SnackbarContent
+        style={{
+          backgroundColor: "green",
+        }}
+        message={<span id="client-snackbar">Account updated. 👍</span>}
+      />
+    </Snackbar>
+  );
+}
+
+export default Snack;
